+++ skeleton_replay/interface/__init__.py
@@ -2,7 +2,7 @@
 
 from skeleton_replay.interface.artifacts import ArtifactGenerationPipeline, ArtifactGenerationResult, ArtifactPaths
 from skeleton_replay.interface.console import ColorMode, SkeletonConsole
-from skeleton_replay.interface.output_paths import OutputPathResolver
+from skeleton_replay.interface.output_paths import OutputPathResolver, PytestOutputPathResolver
 from skeleton_replay.interface.report_opener import HtmlReportOpener
 
 __all__ = [
@@ -12,5 +12,6 @@ __all__ = [
     "ColorMode",
     "HtmlReportOpener",
     "OutputPathResolver",
+    "PytestOutputPathResolver",
     "SkeletonConsole",
 ]